Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydub
python-dotenv
sounddevice 
soundfile
requests